+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Leading Systems Contao Merconis bundle changelog
 ===========================================
 
+### 5.1.7 (2025-08-01)
+ * fix scroll to top-pagination
+ * fix sitemap output
+ * fix default Product Page
+ * rename dynamicAttachmentPdfPaths to dynamicAttachmentPaths
+
 ### 5.1.6 (2025-07-18)
  * improve customizer public storage
  * improve UX 'Shopping cart button'